+++ src/main/java/net/silentchaos512/gear/api/item/GearType.java
@@ -64,7 +64,8 @@ public final class GearType {
             b.toolActions(ToolActions.AXE_DIG, ToolActions.PICKAXE_DIG, ToolActions.SHOVEL_DIG));
     public static final GearType PROSPECTOR_HAMMER = getOrCreate("prospector_hammer", PICKAXE, b ->
             b.toolActions(ToolActions.DEFAULT_PICKAXE_ACTIONS));
-    public static final GearType SHEARS = getOrCreate("shears", HARVEST_TOOL);
+    public static final GearType SHEARS = getOrCreate("shears", HARVEST_TOOL, b ->
+            b.toolActions(ToolActions.DEFAULT_SHEARS_ACTIONS));
     public static final GearType SICKLE = getOrCreate("sickle", HARVEST_TOOL, b ->
             b.toolActions(ToolActions.HOE_DIG));
     // Melee weapons (swords)